Add flag to generate a full, no-reflective implementation of messages.

+++ wire-runtime/src/main/java/com/squareup/wire/TagMap.java
@@ -188,7 +188,7 @@ public final class TagMap {
         try {
           list.add(adapter(targetExtension).decode(reader));
         } catch (RuntimeEnumAdapter.EnumConstantNotFoundException e) {
-          list.add(e.value);
+          list.add((long) e.value);
         }
       }
       reader.endMessage(token);
